Pull size logic out of config.go
